Refactor input and output data.
Major refactor to create a data access layer between all
inputs and outputs from the route optimization tool

Main changes:
1. Introduces config_manager that loads all inputs
2. Introduces intermediary data output objects that store unserialized
output data
3. Introduces output serialize and persistence methods to centralize all
file persistence.
4. Introduces file_manager to manage the names of all output files.
5. Introduce a simple flask api to show how this change allows us
to run the route optimization tool without writing files, capture the
output, and serve that through an API. This API is just a stub for now.

What is left to do:
- Change the cloud context uploader to use the new file manager
- Make it so the matrices persisted in build_tiem_dist_matrix also
follow the new pattern.
- Check if there are any other missed file writing spots

How was this tested?
- Tested the main flow and ensured the solution value did not change
and that all visualization outputs looked the same
- Tested the manual edits flow but without any actual changes to the routes

What else needs to be tested?
1. Manual edits mode with actual changes to the routes.
2. Cloud client mode.